move testutil.go to pkg/config/store
resolve the cycle dependency. this will be okay since
those test utilities will be stripped because they are not
used in the real code.
