--- pkg/config/store/testutil/BUILD
@@ -1,13 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "testutil.go",
-    ],
-    deps = [
-        "//pkg/config/store:go_default_library",
-    ],
-)
